функция query  в классе DB, инициализация
classes/class.DB.php -> includes/class.DB.php,
инициализация из config/db.config.json
